Move Response model to new package util.model
